# If applied, this commit will...
# Why is this change needed?
Fix the refresh bug. Originally, after delete, I had to manually refresh:
1. Changed origin fetch to a function.
2. Convert deleteSurvey to async/await.
3. Called new fetchSurvey function in after delete to refresh view

# How does it address the issue?

# Provide links to any relevant tickets, articles or other resources

+++ src/js/survey.js
@@ -1,22 +1,34 @@
 class Survey {
-  constructor(id, title) {
+  constructor(id, title, question1, question2, question3) {
     this.id = id
     this.title = title
+    this.question1 = question1
+    this.question2 = question2
+    this.question3 = question3
     this.renderSurveyList()
   }
 
-  deleteSurvey() {
-    fetch(`http://localhost:3000/surveys/${this.id}`, {
+  async deleteSurvey() {
+    await fetch(`http://localhost:3000/surveys/${this.id}`, {
       method: 'DELETE'
     })
       .then(() => {
         document.getElementById('survey-container')
           .removeChild(document.getElementById(this.id))
       })
-    // TODO - Delete works! but doesn't refresh to opening index list
-    window.setTimeout(() => {
-      window.location.href = "/";
-    }, 1000);
+    fetchSurveys()
+  }
+
+  submitSurvey() {
+    // TODO - Should fetch from model where responses are handled
+    console.log('Submit survey')
+    console.log(this.id)
+    // fetch(`http://localhost:3000/responses/${this.id}`, {
+    //   method: 'POST'
+    // })
+    //   .then(() => {
+    //     console.log('In then')
+    //   })
   }
 
   // Index List of Surveys
@@ -42,7 +54,11 @@ class Survey {
         <h3>${this.title}</h3>
       </div>
       <div class="card__summary">
-        Add questions
+        <form action="/action_page.php">
+          <input type="radio" name="vehicle1" value="Bike"> ${this.question1}<br>
+          <input type="radio" name="vehicle2" value="Car"> ${this.question2}<br>
+          <input type="radio" name="vehicle3" value="Boat"> ${this.question3}<br><br>
+      </form>
       </div>
     </div>
     <button class="card__btn submit">Submit</button>
@@ -76,10 +92,8 @@ class Survey {
     surveyCard.addEventListener('click', e => {
       document.querySelector('survey-card')
       // TODO add submit to write data for survey response
-      if (e.target.className.includes('submit')) this.renderSurveyList(e)
-      if (e.target.className.includes('delete')) {
-        this.deleteSurvey(e)
-      }
+      if (e.target.className.includes('submit')) this.submitSurvey(e)
+      if (e.target.className.includes('delete')) this.deleteSurvey(e)
     })
   }
 }
